+++ runtimes/CHANGELOG.md
@@ -1,13 +1,35 @@
 # Changelog
 
+## [0.2.23] - 2024-10-24
+
+### Added
+
+- Add `platform` value to `Runtime` feature injected in every Server (#244)
+
+## [0.2.22] - 2024-10-24
+
+### Added
+
+- Add encryption to the accessToken in Identity Management protocol
+
+### Changed
+
+- Minor updates to `GetSsoToken` request parameters
+
+### Removed
+
+- Remove unused `onUpdateSsoTokenManagement` in Identity Management protocol
+
 ## [0.2.21] - 2024-10-23
 
 ### Added
+
 - Add `workspace/applyEdit` functionality to the LSP interface
 - Add `writeFile`, `appendFile` and `mkdir` functionality to `workspace.fs`
 - Create `Notification` interface for servers
 
 ### Changed
+
 - Update `@aws/language-server-runtimes-types` dependency from 0.0.6 to 0.0.7
 
 ## [0.2.20] - 2024-10-11